requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
dj-database-url==0.5.0
Django==2.0.13
django-environ==0.4.5
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.1
factory-boy==2.11.1
Faker==1.0.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
more-itertools==6.0.0
pluggy==0.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pytest==4.3.0
pytest-django==3.4.8
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
